Refactor and code cleanup.

+++ ecommerce.api/Managers/ProductDataManager.cs
@@ -1,3 +1,4 @@
+using AutoMapper;
 using ecommerce.api.Classes;
 using ecommerce.api.Data;
 using ecommerce.api.Entities;
@@ -9,9 +10,11 @@ namespace ecommerce.api.Managers;
 public class ProductDataManager
 {
     private readonly EcommerceDbContext _context;
+    private readonly IMapper _mapper;
 
-    public ProductDataManager(EcommerceDbContext context)
+    public ProductDataManager(IMapper mapper, EcommerceDbContext context)
     {
+        _mapper = mapper;
         _context = context;
     }   
     
@@ -49,7 +52,7 @@ public class ProductDataManager
         return categories;
     }
 
-    public async Task<ProductEntity> GetProduct(Guid id)
+    public async Task<ProductEntity?> GetProduct(Guid id)
     {
         var product = await _context.Products
             .Include(p => p.Images)
@@ -61,15 +64,7 @@ public class ProductDataManager
 
     public async Task<ProductEntity> CreateProduct(ProductModel product)
     {
-        var mappedProduct = new ProductEntity()
-        {
-            Id = product.Id,
-            Name = product.Name,
-            Description = product.Description,
-            Category = product.Category,
-            PricePerUnit = product.PricePerUnit,
-            Discount = product.Discount
-        };
+        var mappedProduct = _mapper.Map<ProductEntity>(product);
         
         await _context.Products.AddAsync(mappedProduct);
         await _context.SaveChangesAsync();
@@ -82,23 +77,32 @@ public class ProductDataManager
         var existingProduct = await _context.Products
             .Include(p => p.Images)
             .FirstOrDefaultAsync(p => p.Id == product.Id);
+
+        var images = await _context.Images.Where(i => i.ProductId == product.Id).ToListAsync();
         
         existingProduct.Name = product.Name;
+        existingProduct.Images = images;
         existingProduct.Description = product.Description;
         existingProduct.Category = product.Category;
+        existingProduct.Color = product.Color;
         existingProduct.PricePerUnit = product.PricePerUnit;
         existingProduct.Discount = product.Discount;
+        
+        existingProduct.UpdatedDate = DateTime.UtcNow;
         
         await _context.SaveChangesAsync();
 
         return existingProduct;
     }
 
-    public async void DeleteProduct(Guid id)
+    public async Task DeleteProduct(Guid id)
     {
         var product = await _context.Products.FirstOrDefaultAsync(p => p.Id == id);
 
-        _context.Products.Remove(product);
-        _context.SaveChanges();
+        if (product != null)
+        {
+            _context.Products.Remove(product);
+            await _context.SaveChangesAsync();
+        }
     }
 }
